pyproject.toml
[project]
name = "django-vue-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-extensions>=3.2.3",
    "djangorestframework>=3.15.2",
    "pillow>=11.1.0",
    "psycopg[binary,pool]>=3.2.4",
    "requests>=2.32.3",
]
